Comparison data

A list of data frames, each describing data that can be used in comparisons. See attr(*, "meta")$Source of each data frame for its source.

Usage

comparison_data

Format

An object of class list of length 13.

Examples

# Metadata for one of the data frames
print( attr(comparison_data[[1]], "meta") )
#>          About
#> 1 Stature (cm)
#>                                                                                                                                                                                                                                                                                                      Source
#> 1 Sperrin, M., Marshall, A. D., Higgins, V., Renehan, A. G., & Buchan, I. E. (2016). Body mass index relates weight to height differently in women and older adults: Serial cross-sectional surveys in England (1992-2011). Journal of Public Health, 38(3), 607-613. https://doi.org/10.1093/pubmed/fdv067
#>                       Notes
#> 1 Reported Cohen's d = 1.97

# Summary of all available data
ret = NULL
for (nn in names(comparison_data)) {
  r1 = data.frame(Label = nn
    , About = strtrim( attr(comparison_data[[nn]], "meta")$About, 25)
    , `Group 1` = comparison_data[[nn]]$group[1]
    , `Group 2` = comparison_data[[nn]]$group[2]
    , Source = strtrim( attr(comparison_data[[nn]], "meta")$Source, 25)
    , check.names = FALSE)
  ret = rbind(ret, r1)
}
print(ret)
#>          Label                     About          Group 1         Group 2
#> 1     stature1              Stature (cm)              Men           Women
#> 2     stature2          Stature (inches)              Men           Women
#> 3    Vitamin.D Increases in 25-hydroxyvi       Vitamin D2      Vitamin D3
#> 4  cholesterol Changes in total choleste         Oat milk   Control drink
#> 5    serotonin Serotonin metabolite (5HI             Male          Female
#> 6       stroke 36-Item Short-Form Health Stroke survivors Normative value
#> 7          IOP Increase in intraocular p  Increase in IOP            <NA>
#> 8          UTI Amoxicillin / clavulanic         Amox-clav     Amoxicillin
#> 9          DVT Heparin for the preventio          Heparin         Placebo
#> 10          MI Aspirin for the preventio          Aspirin         Placebo
#> 11    drug.use          Illicit drug use         12 to 17        18 to 25
#> 12          TB Streptomycin for tubercul     Streptomycin        Bed rest
#> 13     tonsils Size of tonsils of carrie      Noncarriers        Carriers
#>                       Source
#> 1  Sperrin, M., Marshall, A.
#> 2  McGraw, K. O., & Wong, S.
#> 3  Trang, H. M., Cole, D. E.
#> 4  Onning, G., Wallmark, A.,
#> 5  Young, S. N., Gauthier, S
#> 6  Cerniauskaite, M., Quinta
#> 7  Vieira, G. M., Oliveira, 
#> 8  Martinelli, R., Lopes, A.
#> 9  Dechavanne, M., Ville, D.
#> 10 Rosnow, R. L., & Rosentha
#> 11 Substance Abuse and Menta
#> 12 Streptomycin in Tuberculo
#> 13 Agresti, A. (1980). Gener