Dominance effect sizes for continuous variables

Calculate dominance-based effect sizes for continuous (normally distributed) variables.

Usage

d_cont(x, meta = NULL)

Arguments

x

data. See details.

meta

either NULL or a named list.

Value

Estimates with verbose explanations.

Details

Calculate the following:

dominance (aka common language effect size);

stochastic difference (aka Glass rank-biserial correlation);

odds, which, for continuous variables, is a special case of generalized odds ratio; and

standardized mean difference (aka Cohen's d).

For continuous variables, the following effect sizes simplify to one of the above-mentioned effect sizes. Thus, they are not reported separately.

measure of stochastic superiority (aka Vargha and Delaney's A), which simplifies to dominance;

adjusted dominance, which simplifies to dominance;

adjusted stochastic difference (aka Goodman and Kruskal's gamma), which simplifies to stochastic difference; and

win odds, which simplifies to odds.

x can contain either known parameter values, summary statistics, or microdata.

If x contains known parameter values or summary statistics, it should be a data.frame with 2 rows. It can optionally contain a variable called group, indicating the name of the group. In addition, it must contain one of the following:

expected value (mu) and standard deviation (sigma), which are known parameters; or

sample mean (xbar), sample standard deviation (sd), and sample size (n); or

xbar, standard error of the mean (sem), and n.

sd is the "sample" standard deviation, with denominator n - 1. This is what is returned by the sd() function.

If x contains microdata, it should be a named list with 2 elements, each being a vector of numeric data.

Optionally, x can have an attribute called "meta", which must be a named list.

See also

Examples

# Known parameter values
d_cont(comparison_data$stature1)
#>                                               Data                                               
#> ┌─────────────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What                │ Value                                                                   │
#> ├─────────────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Input               │ comparison_data$stature1                                                │
#> ├─────────────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Groups              │ Men vs Women                                                            │
#> ├─────────────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Expected value (SD) │ the expected value of the [variable] for Men was 175 (SD = 7.1); the    │
#> │                     │ expected value of the [variable] for Women was 161.6 (SD = 6.5)         │
#> ├─────────────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Data                │ $\mu_1 = 175$, $\sigma_1 = 7.1$; $\mu_2 = 161.6$, $\sigma_2 = 6.5$      │
#> └─────────────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        About                                        
#> ┌────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What   │ Value                                                                   │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ About  │ Stature (cm)                                                            │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Source │ Sperrin, M., Marshall, A. D., Higgins, V., Renehan, A. G., & Buchan, I. │
#> │        │ E. (2016). Body mass index relates weight to height differently in      │
#> │        │ women and older adults: Serial cross-sectional surveys in England       │
#> │        │ (1992-2011). Journal of Public Health, 38(3), 607-613.                  │
#> │        │ https://doi.org/10.1093/pubmed/fdv067                                   │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Notes  │ Reported Cohen's d = 1.97                                               │
#> └────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                  Results: Dominance (D)                                  
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Dominance (D)                                                           │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ D (Men vs. Women) = 91.8%                                               │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The probability of a greater [variable] for Men vs. Women is 91.8%.     │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Common language effect size (CLES); Grissom and Kim's probability of    │
#> │             │ superiority; stress-strength model; and reliability.                    │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                            Results: Stochastic difference (SD)                           
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Stochastic difference (SD)                                              │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SD (Men vs. Women) = 0.84                                               │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The correlation between group (Men vs. Women) and outcome is 0.84.      │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Glass rank-biserial correlation; net benefit; Cliff's delta; simple     │
#> │             │ difference formula. SD = one of the two values of Somers' Delta. |SD| = │
#> │             │ Freeman's theta. For binary variables, SD = risk difference (RD), also  │
#> │             │ known as excess risk, attributable risk, absolute risk increase /       │
#> │             │ reduction (ARI / ARD).                                                  │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                           Results: Generalized odds ratio (GOR)                          
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Generalized odds ratio (GOR)                                            │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ GOR (Men vs. Women) = 11.2                                              │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The odds of a greater [variable] for Men vs. Women is 11.2.             │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Agresti's generalized odds ratio for stochastic dominance; win ratio.   │
#> │             │ For binary variables, GOR = odds ratio (OR). For continuous variables,  │
#> │             │ GOR = odds.                                                             │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                        Results: Standardized mean difference (SMD)                       
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Standardized mean difference (SMD)                                      │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SMD (Men vs. Women) = 1.97                                              │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The difference between the means of Men and Women is 1.97 standard      │
#> │             │ deviations.                                                             │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Cohen's d                                                               │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                     Inference                                     
#> ┌──────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What │ Value                                                                   │
#> ├──────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Side │ The effect is in favor of the [variable] being greater for Men than for │
#> │      │ Women.                                                                  │
#> └──────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Methods                                       
#> ┌─────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What    │ Value                                                                   │
#> ├─────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Methods │ Please include this or similar in your Methods section: We present      │
#> │         │ [effect size name], a dominance-based effect size for continuous        │
#> │         │ (normally distributed) variables, calculated using known values of      │
#> │         │ distribution parameters, along with its 95% confidence interval and     │
#> │         │ p-value. Results were considered to be statistically significant if p < │
#> │         │ 0.05. Data analyses were performed using the R package “es4all”         │
#> │         │ (version 0.9.0).                                                        │
#> └─────────┴─────────────────────────────────────────────────────────────────────────┘
#> 

# Average, standard deviation (SD), and sample size
d_cont(comparison_data$stature2)
#>                                            Data                                           
#> ┌──────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What         │ Value                                                                   │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Input        │ comparison_data$stature2                                                │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Groups       │ Men vs Women                                                            │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Average (SD) │ the average [variable] for Men was 69.7 (SD = 2.8); the average         │
#> │              │ [variable] for Women was 64.3 (SD = 2.6)                                │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Data         │ $\bar{x}_1 = 69.7$, $s_1 = 2.8$, $n_1 = 988$; $\bar{x}_2 = 64.3$, $s_2  │
#> │              │ = 2.6$, $n_2 = 1066$                                                    │
#> └──────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        About                                        
#> ┌────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What   │ Value                                                                   │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ About  │ Stature (inches)                                                        │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Source │ McGraw, K. O., & Wong, S. P. (1992). A common language effect size      │
#> │        │ statistic. Psychological Bulletin, 111(2), 361-365.                     │
#> │        │ https://doi.org/10.1037/0033-2909.111.2.361; and Ellis, P. D. (2010).   │
#> │        │ The Essential Guide to Effect Sizes: Statistical Power, Meta-Analysis,  │
#> │        │ and the Interpretation of Research Results. Cambridge University Press. │
#> │        │ https://doi.org/10.1017/CBO9780511761676 (Box 1.3)                      │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Notes  │ Reported D = 92%                                                        │
#> └────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                  Results: Dominance (D)                                  
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Dominance (D)                                                           │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ D (Men vs. Women) = 92.1% (95% CI: 91.0%-93.2%); p-value: < .001        │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The probability of a greater [variable] for Men vs. Women is 92.1%.     │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Common language effect size (CLES); Grissom and Kim's probability of    │
#> │             │ superiority; stress-strength model; and reliability.                    │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                            Results: Stochastic difference (SD)                           
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Stochastic difference (SD)                                              │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SD (Men vs. Women) = 0.84 (95% CI: 0.82-0.86); p-value: < .001          │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The correlation between group (Men vs. Women) and outcome is 0.84.      │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Glass rank-biserial correlation; net benefit; Cliff's delta; simple     │
#> │             │ difference formula. SD = one of the two values of Somers' Delta. |SD| = │
#> │             │ Freeman's theta. For binary variables, SD = risk difference (RD), also  │
#> │             │ known as excess risk, attributable risk, absolute risk increase /       │
#> │             │ reduction (ARI / ARD).                                                  │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                           Results: Generalized odds ratio (GOR)                          
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Generalized odds ratio (GOR)                                            │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ GOR (Men vs. Women) = 11.7 (95% CI: 10.2-13.7); p-value: < .001         │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The odds of a greater [variable] for Men vs. Women is 11.7.             │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Agresti's generalized odds ratio for stochastic dominance; win ratio.   │
#> │             │ For binary variables, GOR = odds ratio (OR). For continuous variables,  │
#> │             │ GOR = odds.                                                             │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                        Results: Standardized mean difference (SMD)                       
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Standardized mean difference (SMD)                                      │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SMD (Men vs. Women) = 2 (95% CI: 1.9-2.11); p-value: < .001             │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The difference between the means of Men and Women is 2 standard         │
#> │             │ deviations.                                                             │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Cohen's d                                                               │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Inference                                       
#> ┌───────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What      │ Value                                                                   │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Inference │ The result **is** statistically significant (p-value: < .001 < 0.05).   │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Side      │ The effect is in favor of the [variable] being greater for Men than for │
#> │           │ Women.                                                                  │
#> └───────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Methods                                       
#> ┌─────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What    │ Value                                                                   │
#> ├─────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Methods │ Please include this or similar in your Methods section: We present      │
#> │         │ [effect size name], a dominance-based effect size for continuous        │
#> │         │ (normally distributed) variables, calculated using maximum likelihood   │
#> │         │ estimation (MLE), along with its 95% confidence interval and p-value.   │
#> │         │ Results were considered to be statistically significant if p < 0.05.    │
#> │         │ Data analyses were performed using the R package “es4all” (version      │
#> │         │ 0.9.0).                                                                 │
#> └─────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
d_cont(comparison_data$Vitamin.D)
#>                                            Data                                           
#> ┌──────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What         │ Value                                                                   │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Input        │ comparison_data$Vitamin.D                                               │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Groups       │ Vitamin D2 vs Vitamin D3                                                │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Average (SD) │ the average [variable] for Vitamin D2 was 13.7 (SD = 11.4); the average │
#> │              │ [variable] for Vitamin D3 was 23.3 (SD = 15.7)                          │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Data         │ $\bar{x}_1 = 13.7$, $s_1 = 11.4$, $n_1 = 17$; $\bar{x}_2 = 23.3$, $s_2  │
#> │              │ = 15.7$, $n_2 = 55$                                                     │
#> └──────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        About                                        
#> ┌────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What   │ Value                                                                   │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ About  │ Increases in 25-hydroxyvitamin D concentrations due to vitamin D        │
#> │        │ supplementation (nmol/L)                                                │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Source │ Trang, H. M., Cole, D. E., Rubin, L. A., Pierratos, A., Siu, S., &      │
#> │        │ Vieth, R. (1998). Evidence that vitamin D3 increases serum              │
#> │        │ 25-hydroxyvitamin D more efficiently than does vitamin D2. The American │
#> │        │ Journal of Clinical Nutrition, 68(4), 854-858.                          │
#> │        │ https://doi.org/10.1093/ajcn/68.4.854                                   │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Notes  │ Reported p-value: 0.03                                                  │
#> └────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                  Results: Dominance (D)                                  
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Dominance (D)                                                           │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ D (Vitamin D3 vs. Vitamin D2) = 69.3% (95% CI: 55.9%-82.4%); p-value:   │
#> │             │ 0.004                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The probability of a greater [variable] for Vitamin D3 vs. Vitamin D2   │
#> │             │ is 69.3%.                                                               │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Common language effect size (CLES); Grissom and Kim's probability of    │
#> │             │ superiority; stress-strength model; and reliability.                    │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                            Results: Stochastic difference (SD)                           
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Stochastic difference (SD)                                              │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SD (Vitamin D3 vs. Vitamin D2) = 0.39 (95% CI: 0.12-0.65); p-value:     │
#> │             │ 0.004                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The correlation between group (Vitamin D3 vs. Vitamin D2) and outcome   │
#> │             │ is 0.39.                                                                │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Glass rank-biserial correlation; net benefit; Cliff's delta; simple     │
#> │             │ difference formula. SD = one of the two values of Somers' Delta. |SD| = │
#> │             │ Freeman's theta. For binary variables, SD = risk difference (RD), also  │
#> │             │ known as excess risk, attributable risk, absolute risk increase /       │
#> │             │ reduction (ARI / ARD).                                                  │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                           Results: Generalized odds ratio (GOR)                          
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Generalized odds ratio (GOR)                                            │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ GOR (Vitamin D3 vs. Vitamin D2) = 2.3 (95% CI: 1.3-4.7); p-value: 0.004 │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The odds of a greater [variable] for Vitamin D3 vs. Vitamin D2 is 2.3.  │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Agresti's generalized odds ratio for stochastic dominance; win ratio.   │
#> │             │ For binary variables, GOR = odds ratio (OR). For continuous variables,  │
#> │             │ GOR = odds.                                                             │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                        Results: Standardized mean difference (SMD)                       
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Standardized mean difference (SMD)                                      │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SMD (Vitamin D3 vs. Vitamin D2) = 0.71 (95% CI: 0.21-1.31); p-value:    │
#> │             │ 0.004                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The difference between the means of Vitamin D3 and Vitamin D2 is 0.71   │
#> │             │ standard deviations.                                                    │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Cohen's d                                                               │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Inference                                       
#> ┌───────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What      │ Value                                                                   │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Inference │ The result **is** statistically significant (p-value: 0.004 < 0.05).    │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Side      │ The effect is in favor of the [variable] being greater for Vitamin D3   │
#> │           │ than for Vitamin D2.                                                    │
#> └───────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Methods                                       
#> ┌─────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What    │ Value                                                                   │
#> ├─────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Methods │ Please include this or similar in your Methods section: We present      │
#> │         │ [effect size name], a dominance-based effect size for continuous        │
#> │         │ (normally distributed) variables, calculated using maximum likelihood   │
#> │         │ estimation (MLE), along with its 95% confidence interval and p-value.   │
#> │         │ Results were considered to be statistically significant if p < 0.05.    │
#> │         │ Data analyses were performed using the R package “es4all” (version      │
#> │         │ 0.9.0).                                                                 │
#> └─────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
d_cont(comparison_data$cholesterol)
#>                                            Data                                           
#> ┌──────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What         │ Value                                                                   │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Input        │ comparison_data$cholesterol                                             │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Groups       │ Oat milk vs Control drink                                               │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Average (SD) │ the average [variable] for Oat milk was -0.17 (SD = 0.59); the average  │
#> │              │ [variable] for Control drink was 0.19 (SD = 0.54)                       │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Data         │ $\bar{x}_1 = -0.17$, $s_1 = 0.59$, $n_1 = 29$; $\bar{x}_2 = 0.19$, $s_2 │
#> │              │ = 0.54$, $n_2 = 23$                                                     │
#> └──────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        About                                        
#> ┌────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What   │ Value                                                                   │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ About  │ Changes in total cholesterol due to an oat milk drink (mmol/l)          │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Source │ Onning, G., Wallmark, A., Persson, M., Akesson, B., Elmstahl, S., &     │
#> │        │ Oste, R. (1999). Consumption of oat milk for 5 weeks lowers serum       │
#> │        │ cholesterol and LDL cholesterol in free-living men with moderate        │
#> │        │ hypercholesterolemia. Annals of Nutrition & Metabolism, 43(5), 301-309. │
#> │        │ https://doi.org/10.1159/000012798                                       │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Notes  │ Reported p-value: 0.005                                                 │
#> └────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                  Results: Dominance (D)                                  
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Dominance (D)                                                           │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ D (Control drink vs. Oat milk) = 67.7% (95% CI: 52.9%-81.9%); p-value:  │
#> │             │ 0.016                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The probability of a greater [variable] for Control drink vs. Oat milk  │
#> │             │ is 67.7%.                                                               │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Common language effect size (CLES); Grissom and Kim's probability of    │
#> │             │ superiority; stress-strength model; and reliability.                    │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                            Results: Stochastic difference (SD)                           
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Stochastic difference (SD)                                              │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SD (Control drink vs. Oat milk) = 0.35 (95% CI: 0.06-0.64); p-value:    │
#> │             │ 0.016                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The correlation between group (Control drink vs. Oat milk) and outcome  │
#> │             │ is 0.35.                                                                │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Glass rank-biserial correlation; net benefit; Cliff's delta; simple     │
#> │             │ difference formula. SD = one of the two values of Somers' Delta. |SD| = │
#> │             │ Freeman's theta. For binary variables, SD = risk difference (RD), also  │
#> │             │ known as excess risk, attributable risk, absolute risk increase /       │
#> │             │ reduction (ARI / ARD).                                                  │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                           Results: Generalized odds ratio (GOR)                          
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Generalized odds ratio (GOR)                                            │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ GOR (Control drink vs. Oat milk) = 2.1 (95% CI: 1.1-4.5); p-value:      │
#> │             │ 0.016                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The odds of a greater [variable] for Control drink vs. Oat milk is 2.1. │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Agresti's generalized odds ratio for stochastic dominance; win ratio.   │
#> │             │ For binary variables, GOR = odds ratio (OR). For continuous variables,  │
#> │             │ GOR = odds.                                                             │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                        Results: Standardized mean difference (SMD)                       
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Standardized mean difference (SMD)                                      │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SMD (Control drink vs. Oat milk) = 0.65 (95% CI: 0.1-1.29); p-value:    │
#> │             │ 0.016                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The difference between the means of Control drink and Oat milk is 0.65  │
#> │             │ standard deviations.                                                    │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Cohen's d                                                               │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Inference                                       
#> ┌───────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What      │ Value                                                                   │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Inference │ The result **is** statistically significant (p-value: 0.016 < 0.05).    │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Side      │ The effect is in favor of the [variable] being greater for Control      │
#> │           │ drink than for Oat milk.                                                │
#> └───────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Methods                                       
#> ┌─────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What    │ Value                                                                   │
#> ├─────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Methods │ Please include this or similar in your Methods section: We present      │
#> │         │ [effect size name], a dominance-based effect size for continuous        │
#> │         │ (normally distributed) variables, calculated using maximum likelihood   │
#> │         │ estimation (MLE), along with its 95% confidence interval and p-value.   │
#> │         │ Results were considered to be statistically significant if p < 0.05.    │
#> │         │ Data analyses were performed using the R package “es4all” (version      │
#> │         │ 0.9.0).                                                                 │
#> └─────────┴─────────────────────────────────────────────────────────────────────────┘
#> 

# Average, standard error of the mean (SEM), and sample size
d_cont(comparison_data$serotonin)
#>                                            Data                                            
#> ┌───────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What          │ Value                                                                   │
#> ├───────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Input         │ comparison_data$serotonin                                               │
#> ├───────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Groups        │ Male vs Female                                                          │
#> ├───────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Average (SEM) │ the average [variable] for Male was 25.7 (SEM = 1.5); the average       │
#> │               │ [variable] for Female was 32.4 (SEM = 2.4)                              │
#> ├───────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Data          │ $\bar{x}_1 = 25.7$, $s_{\bar{x},1} = 1.5$, $n_1 = 43$; $\bar{x}_2 =     │
#> │               │ 32.4$, $s_{\bar{x},2} = 2.4$, $n_2 = 29$                                │
#> └───────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        About                                        
#> ┌────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What   │ Value                                                                   │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ About  │ Serotonin metabolite (5HIAA) concentration in cisternal CSF (ng/ml)     │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Source │ Young, S. N., Gauthier, S., Anderson, G. M., & Purdy, W. C. (1980).     │
#> │        │ Tryptophan, 5-hydroxyindoleacetic acid and indoleacetic acid in human   │
#> │        │ cerebrospinal fluid: Interrelationships and the influence of age, sex,  │
#> │        │ epilepsy and anticonvulsant drugs. Journal of Neurology, Neurosurgery,  │
#> │        │ and Psychiatry, 43(5), 438-445. https://doi.org/10.1136/jnnp.43.5.438   │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Notes  │ Reported p-value < 0.05. Reported Cohen's d = 0.58.                     │
#> └────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                  Results: Dominance (D)                                  
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Dominance (D)                                                           │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ D (Female vs. Male) = 66.0% (95% CI: 52.3%-78.3%); p-value: 0.016       │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The probability of a greater [variable] for Female vs. Male is 66.0%.   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Common language effect size (CLES); Grissom and Kim's probability of    │
#> │             │ superiority; stress-strength model; and reliability.                    │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                            Results: Stochastic difference (SD)                           
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Stochastic difference (SD)                                              │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SD (Female vs. Male) = 0.32 (95% CI: 0.05-0.57); p-value: 0.016         │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The correlation between group (Female vs. Male) and outcome is 0.32.    │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Glass rank-biserial correlation; net benefit; Cliff's delta; simple     │
#> │             │ difference formula. SD = one of the two values of Somers' Delta. |SD| = │
#> │             │ Freeman's theta. For binary variables, SD = risk difference (RD), also  │
#> │             │ known as excess risk, attributable risk, absolute risk increase /       │
#> │             │ reduction (ARI / ARD).                                                  │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                           Results: Generalized odds ratio (GOR)                          
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Generalized odds ratio (GOR)                                            │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ GOR (Female vs. Male) = 1.9 (95% CI: 1.1-3.6); p-value: 0.016           │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The odds of a greater [variable] for Female vs. Male is 1.9.            │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Agresti's generalized odds ratio for stochastic dominance; win ratio.   │
#> │             │ For binary variables, GOR = odds ratio (OR). For continuous variables,  │
#> │             │ GOR = odds.                                                             │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                        Results: Standardized mean difference (SMD)                       
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Standardized mean difference (SMD)                                      │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SMD (Female vs. Male) = 0.58 (95% CI: 0.08-1.11); p-value: 0.016        │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The difference between the means of Female and Male is 0.58 standard    │
#> │             │ deviations.                                                             │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Cohen's d                                                               │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Inference                                       
#> ┌───────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What      │ Value                                                                   │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Inference │ The result **is** statistically significant (p-value: 0.016 < 0.05).    │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Side      │ The effect is in favor of the [variable] being greater for Female than  │
#> │           │ for Male.                                                               │
#> └───────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Methods                                       
#> ┌─────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What    │ Value                                                                   │
#> ├─────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Methods │ Please include this or similar in your Methods section: We present      │
#> │         │ [effect size name], a dominance-based effect size for continuous        │
#> │         │ (normally distributed) variables, calculated using maximum likelihood   │
#> │         │ estimation (MLE), along with its 95% confidence interval and p-value.   │
#> │         │ Results were considered to be statistically significant if p < 0.05.    │
#> │         │ Data analyses were performed using the R package “es4all” (version      │
#> │         │ 0.9.0).                                                                 │
#> └─────────┴─────────────────────────────────────────────────────────────────────────┘
#>