Dominance effect sizes for one-sample continuous variables

Calculate dominance-based effect sizes for continuous (normally distributed) variables in the one-sample or paired scenario.

Usage

d_one(x, meta = NULL, .input = NULL)

Arguments

x

data. See details.

meta

either NULL or a named list.

.input

(don't use)

Value

Estimates with verbose explanations.

Details

Calculate the following:

dominance (aka common language effect size);

stochastic difference (aka Glass rank-biserial correlation);

odds, which, for continuous variables, is a special case of generalized odds ratio; and

standardized mean difference (aka Cohen's d).

For continuous variables, the following effect sizes simplify to one of the above-mentioned effect sizes. Thus, they are not reported separately.

measure of stochastic superiority (aka Vargha and Delaney's A), which simplifies to dominance;

adjusted dominance, which simplifies to dominance; and

adjusted stochastic difference (aka Goodman and Kruskal's gamma), which simplifies to stochastic difference; and

win odds, which simplifies to odds.

x can contain either known parameter values, summary statistics, or microdata.

If x contains known parameter values or summary statistics, it should be a data.frame with 2 rows. It can optionally contain a variable called group, indicating the name of the group. In addition, it must contain one of the following:

expected value (mu) and standard deviation (sigma), which are known parameters; or

sample mean (xbar), sample standard deviation (sd), and sample size (n); or

xbar, standard error of the mean (sem), and n.

sd is the "sample" standard deviation, with denominator n - 1. This is what is returned by the sd() function.

One of the rows describes the population. The other row describes the constant. For the row that describes the constant, either mu or xbar is equal to the constant, while the other variables are set to NA.

If x contains microdata, it should be a named list with 2 elements, each being a vector of numeric data. The element describing the constant must have length 1, while the element describing the population must have a length of > 1.

Optionally, x can have an attribute called "meta", which must be a named list.

See also

Examples

d_one(comparison_data$stroke)
#>                                            Data                                           
#> ┌──────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What         │ Value                                                                   │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Input        │ comparison_data$stroke                                                  │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Groups       │ Stroke survivors vs Normative value                                     │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Average (SD) │ the average [variable] for Stroke survivors was 61 (SD = 20.4);         │
#> │              │ constant Normative value was 66.5                                       │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Data         │ $\bar{x} = 61$, $s = 20.4$, $n = 111$, $m = 66.5$                       │
#> └──────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        About                                        
#> ┌────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What   │ Value                                                                   │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ About  │ 36-Item Short-Form Health Survey (SF-36), mental health (MH)            │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Source │ Cerniauskaite, M., Quintas, R., Koutsogeorgou, E., Meucci, P., Sattin,  │
#> │        │ D., Leonardi, M., & Raggi, A. (2012). Quality-of-life and disability in │
#> │        │ patients with stroke. American Journal of Physical Medicine &           │
#> │        │ Rehabilitation, 91(13 Suppl 1), S39-47.                                 │
#> │        │ https://doi.org/10.1097/PHM.0b013e31823d4df7 (Table 1)                  │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Notes  │ Reported p-value: 0.005                                                 │
#> └────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                  Results: Dominance (D)                                  
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Dominance (D)                                                           │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ D (Normative value vs. Stroke survivors) = 60.7% (95% CI: 53.6%-68.4%); │
#> │             │ p-value: 0.005                                                          │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The probability of a greater [variable] for Normative value vs. Stroke  │
#> │             │ survivors is 60.7%.                                                     │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Common language effect size (CLES); Grissom and Kim's probability of    │
#> │             │ superiority; stress-strength model; and reliability.                    │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                            Results: Stochastic difference (SD)                           
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Stochastic difference (SD)                                              │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SD (Normative value vs. Stroke survivors) = 0.21 (95% CI: 0.07-0.37);   │
#> │             │ p-value: 0.005                                                          │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The correlation between group (Normative value vs. Stroke survivors)    │
#> │             │ and outcome is 0.21.                                                    │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Glass rank-biserial correlation; net benefit; Cliff's delta; simple     │
#> │             │ difference formula. SD = one of the two values of Somers' Delta. |SD| = │
#> │             │ Freeman's theta. For binary variables, SD = risk difference (RD), also  │
#> │             │ known as excess risk, attributable risk, absolute risk increase /       │
#> │             │ reduction (ARI / ARD).                                                  │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                           Results: Generalized odds ratio (GOR)                          
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Generalized odds ratio (GOR)                                            │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ GOR (Normative value vs. Stroke survivors) = 1.5 (95% CI: 1.2-2.2);     │
#> │             │ p-value: 0.005                                                          │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The odds of a greater [variable] for Normative value vs. Stroke         │
#> │             │ survivors is 1.5.                                                       │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Agresti's generalized odds ratio for stochastic dominance; win ratio.   │
#> │             │ For binary variables, GOR = odds ratio (OR). For continuous variables,  │
#> │             │ GOR = odds.                                                             │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                        Results: Standardized mean difference (SMD)                       
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Standardized mean difference (SMD)                                      │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SMD (Normative value vs. Stroke survivors) = 0.38 (95% CI: 0.13-0.68);  │
#> │             │ p-value: 0.005                                                          │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The difference between the means of Normative value and Stroke          │
#> │             │ survivors is 0.38 standard deviations.                                  │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Cohen's d                                                               │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Inference                                       
#> ┌───────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What      │ Value                                                                   │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Inference │ The result **is** statistically significant (p-value: 0.005 < 0.05).    │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Side      │ The effect is in favor of the [variable] being greater for Normative    │
#> │           │ value than for Stroke survivors.                                        │
#> └───────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Methods                                       
#> ┌─────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What    │ Value                                                                   │
#> ├─────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Methods │ Please include this or similar in your Methods section: We present      │
#> │         │ [effect size name], a dominance-based effect size for continuous        │
#> │         │ (normally distributed) variables in the one-sample or paired scenario,  │
#> │         │ calculated using maximum likelihood estimation (MLE), along with its    │
#> │         │ 95% confidence interval and p-value. Results were considered to be      │
#> │         │ statistically significant if p < 0.05. Data analyses were performed     │
#> │         │ using the R package “es4all” (version 0.9.0).                           │
#> └─────────┴─────────────────────────────────────────────────────────────────────────┘
#> 

# Paired data. Compare with `d_pair()`.
df1 = rbind(comparison_data$IOP, data.frame(group = "Zero", xbar = 0, sd = NA, n = NA))
d_one(df1)
#>                                            Data                                           
#> ┌──────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What         │ Value                                                                   │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Input        │ df1                                                                     │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Groups       │ Increase in IOP vs Zero                                                 │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Average (SD) │ the average [variable] for Increase in IOP was 2.2 (SD = 3); constant   │
#> │              │ Zero was 0                                                              │
#> ├──────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Data         │ $\bar{x} = 2.2$, $s = 3$, $n = 29$, $m = 0$                             │
#> └──────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        About                                        
#> ┌────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What   │ Value                                                                   │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ About  │ Increase in intraocular pressure (IOP) (mm Hg) due to weightlifting     │
#> │        │ (mode II)                                                               │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Source │ Vieira, G. M., Oliveira, H. B., de Andrade, D. T., Bottaro, M., &       │
#> │        │ Ritch, R. (2006). Intraocular pressure variation during weight lifting. │
#> │        │ Archives of Ophthalmology (Chicago, Ill.: 1960), 124(9), 1251-1254.     │
#> │        │ https://doi.org/10.1001/archopht.124.9.1251 (Figure 3)                  │
#> ├────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Notes  │ Reported p-value: < .001. IOP increased in 62% of subjects. IOP         │
#> │        │ increased by > 5.0 in 21%.                                              │
#> └────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                  Results: Dominance (D)                                  
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Dominance (D)                                                           │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ D (Increase in IOP vs. Zero) = 77.2% (95% CI: 64.8%-89.6%); p-value: <  │
#> │             │ .001                                                                    │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The probability of a greater [variable] for Increase in IOP vs. Zero is │
#> │             │ 77.2%.                                                                  │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Common language effect size (CLES); Grissom and Kim's probability of    │
#> │             │ superiority; stress-strength model; and reliability.                    │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                            Results: Stochastic difference (SD)                           
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Stochastic difference (SD)                                              │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SD (Increase in IOP vs. Zero) = 0.54 (95% CI: 0.3-0.79); p-value: <     │
#> │             │ .001                                                                    │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The correlation between group (Increase in IOP vs. Zero) and outcome is │
#> │             │ 0.54.                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Glass rank-biserial correlation; net benefit; Cliff's delta; simple     │
#> │             │ difference formula. SD = one of the two values of Somers' Delta. |SD| = │
#> │             │ Freeman's theta. For binary variables, SD = risk difference (RD), also  │
#> │             │ known as excess risk, attributable risk, absolute risk increase /       │
#> │             │ reduction (ARI / ARD).                                                  │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                           Results: Generalized odds ratio (GOR)                          
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Generalized odds ratio (GOR)                                            │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ GOR (Increase in IOP vs. Zero) = 3.4 (95% CI: 1.8-8.6); p-value: < .001 │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The odds of a greater [variable] for Increase in IOP vs. Zero is 3.4.   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Agresti's generalized odds ratio for stochastic dominance; win ratio.   │
#> │             │ For binary variables, GOR = odds ratio (OR). For continuous variables,  │
#> │             │ GOR = odds.                                                             │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                        Results: Standardized mean difference (SMD)                       
#> ┌─────────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What        │ Value                                                                   │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Effect size │ Standardized mean difference (SMD)                                      │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Result      │ SMD (Increase in IOP vs. Zero) = 1.06 (95% CI: 0.54-1.78); p-value: <   │
#> │             │ .001                                                                    │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Meaning     │ The difference between the means of Increase in IOP and Zero is 1.06    │
#> │             │ standard deviations.                                                    │
#> ├─────────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Other names │ Cohen's d                                                               │
#> └─────────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Inference                                       
#> ┌───────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What      │ Value                                                                   │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Inference │ The result **is** statistically significant (p-value: < .001 < 0.05).   │
#> ├───────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Side      │ The effect is in favor of the [variable] being greater for Increase in  │
#> │           │ IOP than for Zero.                                                      │
#> └───────────┴─────────────────────────────────────────────────────────────────────────┘
#> 
#>                                        Methods                                       
#> ┌─────────┬─────────────────────────────────────────────────────────────────────────┐
#> │ What    │ Value                                                                   │
#> ├─────────┼─────────────────────────────────────────────────────────────────────────┤
#> │ Methods │ Please include this or similar in your Methods section: We present      │
#> │         │ [effect size name], a dominance-based effect size for continuous        │
#> │         │ (normally distributed) variables in the one-sample or paired scenario,  │
#> │         │ calculated using maximum likelihood estimation (MLE), along with its    │
#> │         │ 95% confidence interval and p-value. Results were considered to be      │
#> │         │ statistically significant if p < 0.05. Data analyses were performed     │
#> │         │ using the R package “es4all” (version 0.9.0).                           │
#> └─────────┴─────────────────────────────────────────────────────────────────────────┘
#>