Estimator properties of dominance-based effect sizes

Calculate the root mean squared error (RMSE), coverage, power, and false positive rate (FPR) of dominance-based effect size estimators – the maximum likelihood estimator (MLE) and the non-parametric estimator (NPE). Compares power and FPR with that of certain conventional tests.

Warning. This function takes a very long time to run as it uses simulations.

Usage

estimator_properties(
  x,
  nsim,
  seed = 42,
  Sys.sleep_time = if (nsim > 100) 5 else 0
)

Arguments

x

a 2-row data.frame with simulation parameters

nsim

number of simulations

seed

random number seed (ignored if NA)

Sys.sleep_time

occasional calls to Sys.sleep()

Value

Estimator properties.

Details

x is a data.frame with 2 rows that specify how data is generated. It has a variable called n (sample size). In addition, it has the following variables:

Continuous (normally distributed) data: mu and sigma.

Binary data: p.

Examples

# \donttest{
# **Warning.** This function takes a very long time to run as it uses simulations.
# In a real run, update the value of `nsim`.
# These examples have a theoretical power of 80% and a theoretical FPR of 5%.
nsim = 2 ## 5000
estimator_properties(data.frame(mu = c(105, 100), sigma = c(10, 10), n = c(64, 64)), nsim = nsim)
#>                                         Parameters                                        
#> ┌────────────────────────────────────────────────────────────────────────────────────────┐
#> │ Value                                                                                  │
#> ├────────────────────────────────────────────────────────────────────────────────────────┤
#> │ $\mu_1 = 105$, $\sigma_1 = 10$, $n_1 = 64$; $\mu_2 = 100$, $\sigma_2 = 10$, $n_2 = 64$ │
#> └────────────────────────────────────────────────────────────────────────────────────────┘
#> 
#>                           RMSE and coverage                           
#> ┌────────────────────────────────────────────────────────────────────┐
#> │ Value                                                              │
#> ├────────────────────────────────────────────────────────────────────┤
#> │ | SD | RMSE (MLE) | RMSE (NPE) | Coverage (MLE) | Coverage (NPE) | │
#> │ |---|---|---|---|---|                                              │
#> │ | 0.276 | 0.034 | 0.061 | 100.0% | 100.0% |                        │
#> └────────────────────────────────────────────────────────────────────┘
#> 
#>                       Power                      
#> ┌───────────────────────────────────────────────┐
#> │ Value                                         │
#> ├───────────────────────────────────────────────┤
#> │ | MLE | NPE | Welch's t | Student's t | WMW | │
#> │ |---|---|---|---|---|                         │
#> │ |100.0%|100.0%|100.0%|100.0%|100.0%|          │
#> └───────────────────────────────────────────────┘
#> 
#>                        FPR                       
#> ┌───────────────────────────────────────────────┐
#> │ Value                                         │
#> ├───────────────────────────────────────────────┤
#> │ | MLE | NPE | Welch's t | Student's t | WMW | │
#> │ |---|---|---|---|---|                         │
#> │ |0.0%|0.0%|0.0%|0.0%|0.0%|                    │
#> └───────────────────────────────────────────────┘
#> 
estimator_properties(data.frame(p = c(0.6, 0.4), n = c(97, 97)), nsim = nsim)
#>                        Parameters                       
#> ┌──────────────────────────────────────────────────────┐
#> │ Value                                                │
#> ├──────────────────────────────────────────────────────┤
#> │ $\pi_1 = 0.6$, $n_1 = 97$; $\pi_2 = 0.4$, $n_2 = 97$ │
#> └──────────────────────────────────────────────────────┘
#> 
#>                           RMSE and coverage                           
#> ┌────────────────────────────────────────────────────────────────────┐
#> │ Value                                                              │
#> ├────────────────────────────────────────────────────────────────────┤
#> │ | SD | RMSE (MLE) | RMSE (NPE) | Coverage (MLE) | Coverage (NPE) | │
#> │ |---|---|---|---|---|                                              │
#> │ | 0.2 | 0.123 | 0.123 | 100.0% | 100.0% |                          │
#> └────────────────────────────────────────────────────────────────────┘
#> 
#>                           Power                           
#> ┌────────────────────────────────────────────────────────┐
#> │ Value                                                  │
#> ├────────────────────────────────────────────────────────┤
#> │ | MLE | NPE | Fisher's exact | Pearson's chi-squared | │
#> │ |---|---|---|---|                                      │
#> │ |0.0%|0.0%|0.0%|0.0%|                                  │
#> └────────────────────────────────────────────────────────┘
#> 
#>                            FPR                            
#> ┌────────────────────────────────────────────────────────┐
#> │ Value                                                  │
#> ├────────────────────────────────────────────────────────┤
#> │ | MLE | NPE | Fisher's exact | Pearson's chi-squared | │
#> │ |---|---|---|---|                                      │
#> │ |50.0%|50.0%|50.0%|50.0%|                              │
#> └────────────────────────────────────────────────────────┘
#> 
# }