Set certain options

set_opts() sets certain package options. To view these options, use show_opts(). For more advanced control and detailed customization, experienced users can also employ options() and show_options() (refer to es4all-options for further information).

Usage

set_opts(
  reset = NULL,
  conf.level = NULL,
  es = NULL,
  output = NULL,
  file = NULL,
  .file_temp = NULL,
  raw = NULL
)

show_opts()

Arguments

reset

reset all options to their default values?

conf.level

confidence level of the interval.

es

which effect sizes to print? NA (the default, prints all), or a vector of short effect size names. See details.

output

specify how the output is printed: "auto" (default); "huxtable", "gt", "kableExtra", "flextable"; "screen". For the following output types, please also specify the file argument: "Excel", "Word", "CSV".

file

file name (see output).

.file_temp

place file in a temporary folder?

raw

if TRUE, d_*() functions return the raw intermediate data.frame instead of the formatted output. Default: FALSE.

Value

(Nothing.)

Details

If you are not setting a particular option, leave it as NULL.

If es = NA, print all effect sizes. If es is a vector of short effect size names, print just those effect sizes. Possible values are: "SD", "ASD", "NNT", "D", "MSS", "AD", "GOR", "WO", "SMD".

output determines how the output is printed:

"auto" (default): automatically select the table-making package, depending on the destination (such as screen, HTML, or PDF / LaTeX).

"huxtable", "gt", "kableExtra", "flextable": use this table-making package. Be sure that this package is installed.

"screen": print plain-text tables to the screen.

"Excel": print to an Excel workbook. Please specify the name of an Excel file using the file argument. Before using Excel printing, please be sure to install the openxlsx2 package.

"Word": print to a Word document. Please specify the name of a Word file using the file argument. Before using Word printing, please be sure to install these packages: flextable and officer.

"CSV": print to a comma-separated values (CSV) file. Please specify the name of a CSV file using the file argument.

See also

Examples

set_opts(es = NA)
#> * Effect size: print all effect sizes.